+++ lib/presentation/screens/admin/users_account.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:project_new/presentation/widgets/admin/appbar_admin.dart';
 
 import '../../../data/models/user/user_model.dart';
 import '../../../providers/admin/users_acount_provider.dart';
@@ -59,16 +58,45 @@ class _UsersAccountState extends State<UsersAccount> {
                   trailing: IconButton(
                     icon: Icon(Icons.delete, color: Colors.red),
                     onPressed: () async {
-                      await provider.deleteUser(user.id);
-                      ScaffoldMessenger.of(context).showSnackBar(
-                        SnackBar(
-                          content: Text(
-                            'This user has been successfully deleted',
-                            style: TextStyle(color: Colors.green, fontSize: 15),
-                          ),
-                          duration: Duration(seconds: 2),
-                        ),
+                      final confirm = await showDialog<bool>(
+                        context: context,
+                        builder:
+                            (ctx) => AlertDialog(
+                              title: Text('Confirm Delete'),
+                              content: Text(
+                                'Are you sure you want to delete this user ?',
+                              ),
+                              actions: [
+                                TextButton(
+                                  onPressed: () => Navigator.of(ctx).pop(false),
+                                  child: Text('Cancel'),
+                                ),
+                                TextButton(
+                                  onPressed: () => Navigator.of(ctx).pop(true),
+                                  child: Text(
+                                    'Delete',
+                                    style: TextStyle(color: Colors.red),
+                                  ),
+                                ),
+                              ],
+                            ),
                       );
+
+                      if (confirm == true) {
+                        await provider.deleteUser(user.id);
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          SnackBar(
+                            content: Text(
+                              'User deleted successfully',
+                              style: TextStyle(
+                                color: Colors.green,
+                                fontSize: 15,
+                              ),
+                            ),
+                            duration: Duration(seconds: 2),
+                          ),
+                        );
+                      }
                     },
                   ),
                 ),
